split main headers into smaller ones
