+++ cpp-terminal/old/terminal_base.h
@@ -11,35 +11,6 @@
 
 #include <stdexcept>
 
-#ifdef _WIN32
-#include <conio.h>
-#include <io.h>
-#include <windows.h>
-#else
-#include <sys/ioctl.h>
-#include <termios.h>
-#undef B0
-#undef B50
-#undef B75
-#undef B110
-#undef B134
-#undef B150
-#undef B200
-#undef B300
-#undef B600
-#undef B1200
-#undef B1800
-#undef B2400
-#undef B4800
-#undef B9600
-#undef B19200
-#undef B28800
-#undef B38400
-#undef B57600
-#undef B115200
-#include <unistd.h>
-#include <cerrno>
-#endif
 
 namespace Term {
 
@@ -168,102 +139,9 @@ class BaseTerminal {
 #endif
     }
 
-    // Returns true if a character is read, otherwise immediately returns false
-    bool read_raw(char* s) const {
-        if (!keyboard_enabled) {
-            int c = getchar();
-            if (c >= 0) {
-                *s = c;
-            } else if (c == EOF) {
-                // In non-raw (blocking) mode this happens when the input file
-                // ends. In such a case, return the End of Transmission (EOT)
-                // character (Ctrl-D)
-                *s = 0x04;
-            } else {
-                throw std::runtime_error("getchar() failed");
-            }
-            return true;
-        }
-#ifdef _WIN32
-        char buf[1];
-        DWORD nread;
-        if (_kbhit()) {
-            if (!ReadFile(hin, buf, 1, &nread, nullptr)) {
-                throw std::runtime_error("ReadFile() failed");
-            }
-            if (nread == 1) {
-                *s = buf[0];
-                return true;
-            } else {
-                throw std::runtime_error("kbhit() and ReadFile() inconsistent");
-            }
-        } else {
-            return false;
-        }
-#else
-        int nread = read(STDIN_FILENO, s, 1);
-        if (nread == -1 && errno != EAGAIN) {
-            throw std::runtime_error("read() failed");
-        }
-        return (nread == 1);
-#endif
-    }
 
-    bool get_term_size(int& rows, int& cols) const {
-#ifdef _WIN32
-        CONSOLE_SCREEN_BUFFER_INFO inf;
-        if (GetConsoleScreenBufferInfo(hout, &inf)) {
-            cols = inf.srWindow.Right - inf.srWindow.Left + 1;
-            rows = inf.srWindow.Bottom - inf.srWindow.Top + 1;
-            return true;
-        } else {
-            // This happens when we are not connected to a terminal
-            return false;
-        }
-#else
-        struct winsize ws {};
-        if (ioctl(STDOUT_FILENO, TIOCGWINSZ, &ws) == -1 || ws.ws_col == 0) {
-            // This happens when we are not connected to a terminal
-            return false;
-        } else {
-            cols = ws.ws_col;
-            rows = ws.ws_row;
-            return true;
-        }
-#endif
-    }
 
-    // Returns true if the standard input is attached to a terminal
-    static bool is_stdin_a_tty() {
-#ifdef _WIN32
-        return _isatty(_fileno(stdin));
-#else
-        return isatty(STDIN_FILENO);
-#endif
-    }
 
-    // Returns true if the standard output is attached to a terminal
-    static bool is_stdout_a_tty() {
-#ifdef _WIN32
-        return _isatty(_fileno(stdout));
-#else
-        return isatty(STDOUT_FILENO);
-#endif
-    }
-
-    // coverts a string into an integer
-    static int convert_string_to_int(const char* string,
-                                     const char* format,
-                                     int* rows,
-                                     int* cols) {
-#ifdef _WIN32
-        // windows provides it's own alternative to sscanf()
-        return sscanf_s(string, format, rows, cols);
-#else
-        // TODO move to a better way
-        return sscanf(string, format, rows, cols);
-#endif
-    }
 };
 
 }  // namespace Term
