+++ cpp-terminal/version.hpp.in
