+++ examples/colors.cpp
@@ -1,20 +1,13 @@
-#include <cpp-terminal/terminal.h>
-#include <cpp-terminal/version.h>
+#include <cpp-terminal/base.hpp>
+#include <cpp-terminal/version.hpp>
 
-using Term::bg;
-using Term::color;
-using Term::color24_bg;
-using Term::color24_fg;
-using Term::fg;
-using Term::style;
-using Term::Terminal;
+using namespace Term;
 
 int main() {
     std::cout << "Running cpp-terminal version: "
               << CPP_TERMINAL_VERSION_COMPLETE << std::endl;
     try {
-        Terminal term;
-        if (Terminal::is_stdout_a_tty()) {
+        if (is_stdout_a_tty()) {
             std::cout << "Standard output is attached to a terminal."
                       << std::endl;
         } else {
