+++ examples/menu_window_24bit.cpp
@@ -1,4 +1,5 @@
-#include <cpp-terminal/terminal.h>
+#include <cpp-terminal/input.hpp>
+#include <cpp-terminal/window.hpp>
 
 using Term::bg;
 using Term::fg;
